requirements.txt
fastapi==0.118.3
pydantic==2.12.0
python-dotenv==1.1.1
telethon==1.36.0
